add thumbnail, rename react-marker to html-marker sample
