+++ src/frontend-samples/react-marker-sample/HtmlMarkerTools.tsx
