+++ src/sampleManifest.tsx
@@ -11,7 +11,7 @@ import { getImageExportSpec } from "./frontend-samples/image-export/sampleSpec";
 import { getMarkerPinSpec } from "./frontend-samples/marker-pin-sample/sampleSpec";
 import { getMultiViewportSpec } from "frontend-samples/multi-viewport-sample/sampleSpec";
 import { getPropertyFormattingSpec } from "./frontend-samples/property-formatting-sample/sampleSpec";
-import { getReactMarkerSpec } from "./frontend-samples/react-marker-sample/sampleSpec";
+import { getHtmlMarkerSpec } from "./frontend-samples/react-marker-sample/sampleSpec";
 import { getShadowStudySpec } from "./frontend-samples/shadow-study-sample/sampleSpec";
 import { getThematicDisplaySpec } from "./frontend-samples/thematic-display-sample/sampleSpec";
 import { getSwipingComparisonSpec } from "./frontend-samples/swiping-viewport/sampleSpec";
@@ -85,12 +85,12 @@ export const sampleManifest: SampleSpecGroup[] = [{
     getExplodeSpec(),
     getIotAlertSpec(),
     getHeatmapDecoratorSpec(),
+    getHtmlMarkerSpec(),
     getImageExportSpec(),
     getCrossProbingSpec(),
     getMarkerPinSpec(),
     getMultiViewportSpec(),
     getPropertyFormattingSpec(),
-    getReactMarkerSpec(),
     getReadSettingsSpec(),
     getSerializeViewSpec(),
     getScreenSpaceEffectsSpec(),
